requirements.txt
-f /usr/share/pip-wheels
alabaster==0.7.12
amqp==5.0.1
asgiref==3.2.10
Babel==2.8.0
billiard==3.6.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.0.8
django-anymail==8.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-haystack==3.0
django-infinite-scroll-pagination==1.1.0
django-precise-bbcode==1.2.13
djangorestframework==3.11.1
docutils==0.16
idna==2.10
imagesize==1.2.0
importlib-metadata==2.0.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
mistune==0.8.4
mysqlclient==2.0.1
olefile==0.46
packaging==20.4
Pillow==8.0.1
prompt-toolkit==3.0.7
Pygments==2.7.2
Pynames==0.2.3
pyparsing==2.4.7
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
unicodecsv==0.14.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
Whoosh==2.7.4
zipp==3.2.0
